# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Freedom.AI Operational Hub - API POC Viseu
  version: 0.1.0-poc
  description: >
    Contrato formal da API da POC. Escopo fechado para producao da POC:
    nao implementar alem do escopo contratado e nao deixar de implementar
    nenhuma funcionalidade do escopo.
servers:
  - url: http://localhost:3001
    description: Backend local
...
paths:
...
  /api/v1/trts:
    get:
      tags: [TRTs]
      summary: Listar TRTs
      responses:
        "200":
          description: Lista de TRTs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trt"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
  /api/v1/trts/{id}:
    patch:
      tags: [TRTs]
      summary: Ativar ou desativar TRT
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ativo:
                  type: boolean
              required: [ativo]
      responses:
        "200":
          description: TRT atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trt"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
  /api/v1/webhooks/whatsapp:
    get:
      tags: [Webhooks]
      summary: Verificacao do webhook
      security: []
      parameters:
        - in: query
          name: hub.mode
          schema:
            type: string
        - in: query
          name: hub.verify_token
          schema:
            type: string
        - in: query
          name: hub.challenge
          schema:
            type: string
      responses:
        "200":
          description: Challenge validado
          content:
            text/plain:
              schema:
                type: string
        "403":
          description: Token invalido
    post:
      tags: [Webhooks]
      summary: Receber evento de resposta (n8n/WhatsApp simulado)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Evento recebido
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /api/v1/usuarios:
    get:
      tags: [Usuarios]
      summary: Listar usuarios (admin)
      responses:
        "200":
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Usuario"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
    post:
      tags: [Usuarios]
      summary: Criar usuario (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioCreateRequest"
      responses:
        "201":
          description: Usuario criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "409":
          $ref: "#/components/responses/ConflictResponse"
  /api/v1/usuarios/{id}:
    patch:
      tags: [Usuarios]
      summary: Atualizar usuario (admin)
      parameters:
        - $ref: "#/components/parameters/PathIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UsuarioUpdateRequest"
      responses:
        "200":
          description: Usuario atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Usuario"
        "400":
          $ref: "#/components/responses/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          $ref: "#/components/responses/NotFoundResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathIdParam:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    ValidationErrorResponse:
      description: Erro de validacao
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
    UnauthorizedResponse:
      description: Nao autenticado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ForbiddenResponse:
      description: Sem permissao
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFoundResponse:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ConflictResponse:
      description: Conflito de dados
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    RoleUsuario:
      type: string
      enum: [ADMIN, OPERADOR, GESTOR]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        detalhes:
          type: array
          items:
            type: object
            additionalProperties: true
      required: [error]
...
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nome:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: "#/components/schemas/RoleUsuario"
        ativo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, nome, email, role, ativo]
    UsuarioCreateRequest:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        senha:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/RoleUsuario"
        ativo:
          type: boolean
      required: [nome, email, senha]
    UsuarioUpdateRequest:
      type: object
      properties:
        nome:
          type: string
        email:
          type: string
          format: email
        senha:
          type: string
          minLength: 6
        role:
          $ref: "#/components/schemas/RoleUsuario"
        ativo:
          type: boolean
      additionalProperties: false
...
    Trt:
      type: object
      properties:
        id:
          type: string
          format: uuid
        numero:
          type: string
        nome:
          type: string
        ativo:
          type: boolean
      required: [id, numero, nome, ativo]
...
security:
  - bearerAuth: []